Pulled out common CFamilyStateMachine
