+++ src/matchers/lang/cfamily/States.ts
@@ -1,6 +1,10 @@
+/**
+ * States for C family languages
+ */
 
 import { LangState } from "../LangStateMachine";
+
 export const Normal = new LangState("normal", false, false);
 export const DoubleString = new LangState("string", false, true);
-export const SlashStarComment = new LangState("CComment", true, false);
+export const SlashStarComment = new LangState("/*comment", true, false);
 export const SlashSlashComment = new LangState("//comment", true, false);
